refactor : move file rename and file remove logic from user service to common service
