+++ src/user/user.service.ts
@@ -14,12 +14,14 @@ import { promises } from 'fs';
 import { join } from 'path';
 import { PROFILE_IMAGE_PATH } from 'src/common/const/path.const';
 import * as bcrypt from 'bcryptjs';
+import { CommonService } from 'src/common/common.service';
 
 @Injectable()
 export class UserService {
   constructor(
     @InjectRepository(UserEntity)
     private readonly userRepository: Repository<UserEntity>,
+    private readonly commonService: CommonService,
   ) {}
 
   /**
@@ -120,45 +122,12 @@ export class UserService {
     }
 
     if (userProfile) {
-      user.userProfilePath && this.removeFile(PROFILE_IMAGE_PATH, user.userProfilePath);
+      user.userProfilePath && this.commonService.removeFile(PROFILE_IMAGE_PATH, user.userProfilePath);
       user.userProfilePath = userProfile;
     }
 
     this.userRepository.save(user);
 
     return user;
-  }
-
-  /**
-   * 1. receive current file name with extention
-   * 2. find current profile by received file path and file name
-   * 3. delete current profile
-   */
-  async removeFile(path: string, fileName: string) {
-    const removeFilePath = join(path, fileName);
-
-    await promises.rm(removeFilePath);
-
-    return true;
-  }
-
-  /**
-   * 1. receive current folder and new folder name and file name to move
-   * 2. move file to new folder from current folder
-   */
-  async renameFile(currentPath: string, newPath: string, fileName: string) {
-    const renameFilePath = join(currentPath, fileName);
-
-    try {
-      await promises.access(renameFilePath);
-    } catch (error) {
-      throw new BadRequestException('존재하지 않는 파일입니다.');
-    }
-
-    const newFolderPath = join(newPath, fileName);
-
-    await promises.rename(renameFilePath, newFolderPath);
-
-    return true;
   }
 }
